+++ src/service/firebaseAdmin.ts
@@ -3,7 +3,9 @@ import * as firebaseAdmin from "firebase-admin";
 if (!firebaseAdmin.apps.length) {
   firebaseAdmin.initializeApp({
     credential: firebaseAdmin.credential.cert({
-      privateKey: process.env.NEXT_PUBLIC_PRIVATE_KEY,
+      privateKey: process.env.NEXT_PUBLIC_PRIVATE_KEY
+        ? process.env.NEXT_PUBLIC_PRIVATE_KEY.replace(/\\n/g, "\n")
+        : undefined,
       clientEmail: process.env.NEXT_PUBLIC_CLIENT_EMAIL,
       projectId: process.env.NEXT_PUBLIC_PROJECT_ID,
     }),
